style: show submission status indicator for each task for students
